requirements.txt
quotexpy
python-dotenv
termcolor
python-telegram-bot>=20.0 